+++ app/cells/folio/photoswipe_cell.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-class Folio::PhotoswipeCell < Folio::ApplicationCell
-end